+++ dflib-echarts/src/main/java/org/dflib/echarts/AxisLabel.java
@@ -8,6 +8,11 @@ import org.dflib.echarts.render.option.axis.AxisLabelModel;
 public class AxisLabel {
 
     private String formatter;
+    private Integer rotate;
+    private Integer fontSize;
+    private FontStyle fontStyle;
+    private String fontWeight;
+    private String fontFamily;
 
     /**
      * @since 1.0.0-M22
@@ -33,7 +38,62 @@ public class AxisLabel {
         return this;
     }
 
+    /**
+     * @since 1.0.0-M23
+     */
+    public AxisLabel rotate(int rotateDegrees) {
+        // TODO: validate range -90..90
+        this.rotate = rotateDegrees;
+        return this;
+    }
+
+    /**
+     * @since 1.0.0-M23
+     */
+    public AxisLabel fontSize(int fontSize) {
+        this.fontSize = fontSize;
+        return this;
+    }
+
+    /**
+     * @since 1.0.0-M23
+     */
+    public AxisLabel fontStyle(FontStyle fontStyle) {
+        this.fontStyle = fontStyle;
+        return this;
+    }
+
+    /**
+     * @since 1.0.0-M23
+     */
+    public AxisLabel fontWeight(FontWeight fontWeight) {
+        this.fontWeight = fontWeight != null ? "'" + fontWeight + "'" : null;
+        return this;
+    }
+
+    /**
+     * @since 1.0.0-M23
+     */
+    public AxisLabel fontWeight(int fontWeight) {
+        this.fontWeight = String.valueOf(fontWeight);
+        return this;
+    }
+
+    /**
+     * @since 1.0.0-M23
+     */
+    public AxisLabel fontFamily(String fontFamily) {
+        this.fontFamily = fontFamily;
+        return this;
+    }
+
     protected AxisLabelModel resolve() {
-        return new AxisLabelModel(formatter);
+        return new AxisLabelModel(
+                formatter,
+                rotate,
+                fontSize,
+                fontStyle != null ? fontStyle.name() : null,
+                fontWeight,
+                fontFamily);
     }
 }