+++ dflib-echarts/src/main/java/org/dflib/echarts/FontWeight.java
@@ -1,0 +1,8 @@
+package org.dflib.echarts;
+
+/**
+ * @since 1.0.0-M23
+ */
+public enum FontWeight {
+    normal, bold, bolder, lighter
+}