+++ RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * #311 ColSet.select(RowToValueMapper...) does "merge" instead of "select"
 * #312 DataFrame.get(col, row)
 * #313 Upgrade ECharts to 5.5.1
+* #316 ECharts AxisLabel options: rotate, font...
 
 ## Release 1.0.0-M22